requirements.txt
pandas~=1.4.2
torch~=1.11.0
torchvision~=0.12.0
pillow~=8.4.0
matplotlib~=3.4.3
tqdm~=4.62.3